+++ behat.yml
@@ -10,3 +10,9 @@ default:
             contexts:
                 - FeatureContext
                 - Behat\MinkExtension\Context\MinkContext
+            paths: [%paths.base%/features/web]
+
+        commands:
+            contexts:
+                - CommandLineProcessContext
+            paths: [%paths.base%/features/commands]
