+++ features/bootstrap/FeatureContext.php
@@ -13,8 +13,6 @@ require_once __DIR__. '/../../vendor/phpunit/phpunit/src/Framework/Assert/Functi
  */
 class FeatureContext extends RawMinkContext implements Context
 {
-    private $output;
-
     /**
      * Initializes context.
      *
@@ -24,64 +22,6 @@ class FeatureContext extends RawMinkContext implements Context
      */
     public function __construct()
     {
-    }
-
-    /**
-     * @BeforeScenario
-     */
-    public function moveIntoTestDir()
-    {
-        if (!is_dir('test')) {
-            mkdir('test');
-        }
-        chdir('test');
-    }
-
-    /**
-     * @AfterScenario
-     */
-    public function moveOutOfTestDir()
-    {
-        chdir('..');
-        if (is_dir('test')) {
-            system('rm -r ' . realpath('test'));
-        }
-    }
-
-    /**
-     * @Given there is a file named :filename
-     */
-    public function thereIsAFileNamed($filename)
-    {
-        touch($filename);
-    }
-
-    /**
-     * @When I run :command
-     */
-    public function iRun($command)
-    {
-        $this->output = shell_exec($command);
-    }
-
-    /**
-     * @Then I should see :string in the output
-     */
-    public function iShouldSeeInTheOutput($string)
-    {
-        assertContains(
-            $string,
-            $this->output,
-            sprintf('Did not see "%s" in the output "%s"', $string, $this->output)
-        );
-    }
-
-    /**
-     * @Given there is a dir named :dir
-     */
-    public function thereIsADirNamed($dir)
-    {
-        mkdir($dir);
     }
 
     /**
